frontend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.0
feedparser==6.0.11
requests==2.31.0
django-cors-headers==4.3.1
